Check for directory before trying to remove it

+++ python/Concat_Metadata_CSW_FromDatasource_urlib3.py
@@ -184,7 +184,8 @@ outputDirectory = options.output_directory
 #
 ###############################################################################################################################
 
-shutil.rmtree(outputDirectory)
+if os.path.exists(outputDirectory):
+    shutil.rmtree(outputDirectory)
 print("Constructing directory " + outputDirectory)
 os.makedirs(outputDirectory)
